pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors =["DanyOk32 <danyaverchenko@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.16.1"
mysqlclient = "^2.2.7"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-results = "^2.6.0"
django-celery-beat = "^2.8.1"
djangochannelsrestframework = "^1.3.0"
daphne = "^4.2.1"
channels-redis = "^4.3.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
